refactor(pr): Rename ArtifactSortByMethod to SortStrategy
