+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/artifacts/NpmArtifact.kt
@@ -1,7 +1,7 @@
 package com.netflix.spinnaker.keel.artifacts
 
-import com.netflix.spinnaker.keel.api.artifacts.ArtifactSortByMethod
-import com.netflix.spinnaker.keel.api.artifacts.ArtifactSortByMethod.VERSION
+import com.netflix.spinnaker.keel.api.artifacts.SortStrategy
+import com.netflix.spinnaker.keel.api.artifacts.SortStrategy.VERSION
 import com.netflix.spinnaker.keel.api.artifacts.ArtifactStatus
 import com.netflix.spinnaker.keel.api.artifacts.DeliveryArtifact
 import com.netflix.spinnaker.keel.api.artifacts.NPM
@@ -16,7 +16,7 @@ data class NpmArtifact(
   override val reference: String = name,
   override val statuses: Set<ArtifactStatus> = emptySet(),
   override val versioningStrategy: VersioningStrategy = NpmVersioningStrategy,
-  override val sortBy: ArtifactSortByMethod = VERSION,
+  override val sortBy: SortStrategy = VERSION,
   override val branch: String? = null
 ) : DeliveryArtifact() {
   override val type = NPM
