+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/jackson/mixins/SortStrategyMixin.kt
@@ -2,7 +2,7 @@ package com.netflix.spinnaker.keel.jackson.mixins
 
 import com.fasterxml.jackson.annotation.JsonValue
 
-internal interface ArtifactSortByMethodMixin {
+internal interface SortStrategyMixin {
   @get:JsonValue
   val type: String
 }
